requirements.txt
selenium
pillow
pytesseract
pyautogui
requests

